requirements.txt
PyYAML==6.0
requests==2.28.1
schedule==1.1.0
selenium==4.7.2
webdriver-manager==3.8.5
python-dateutil==2.8.2
